requirements.txt
# Core LiveKit dependencies
livekit>=1.0.8
livekit-agents>=1.1.1

# Plugins for STT, TTS, LLM
livekit-plugins-groq>=1.1.1
livekit-plugins-deepgram>=1.1.1
livekit-plugins-cartesia>=1.1.1
livekit-plugins-silero>=1.1.1

# Environment management
python-dotenv>=1.0.0

# API clients (optional direct API access)
groq>=0.4.0
deepgram-sdk>=3.2.0